feat(audio): fir double precision (64-bit) per request
- FirContext taps/history/gain/preamp -> double, aligned_alloc double
- fir_neon dot: float32x4 vmlaq -> float64x2 vfmaq (2-way, half throughput, ponytail note)
- arm32 fallback scalar double (no NEON f64)
- StreamingDspProcessor firHistory DoubleArray, filtered Double
- measured HK1 arm64: stream 10.6x->5.0x (LOW), 9.2x->4.7x (MED), 8.6x->4.6x (HIGH, realtime 1.34x->0.73x now non-realtime), ponytail: 64-bit costs 2x

+++ audio-dsp-core/src/main/kotlin/com/miruplay/tv/audio/StreamingDspProcessor.kt
@@ -162,7 +162,7 @@ class StreamingDspProcessor(
         private val biquadStates = plan.biquadsByChannel.map { chain ->
             Array(chain.size) { BiquadState() }
         }
-        private val firHistory = plan.firTapsByChannel.map { taps -> FloatArray(taps.size) }
+        private val firHistory = plan.firTapsByChannel.map { taps -> DoubleArray(taps.size) } // ponytail: double precision per user request
         private var firCursor = 0
 
         fun processFrame(input: FloatArray, offset: Int): FloatArray {
@@ -175,10 +175,10 @@ class StreamingDspProcessor(
                 val taps = plan.firTapsByChannel[channel]
                 if (taps.isNotEmpty()) {
                     val history = firHistory[channel]
-                    history[firCursor] = value.toFloat()
+                    history[firCursor] = value
                     var filtered = 0.0
                     for (tap in taps.indices) {
-                        filtered += taps[tap] * history[(firCursor - tap + history.size) % history.size]
+                        filtered += taps[tap].toDouble() * history[(firCursor - tap + history.size) % history.size]
                     }
                     value = filtered
                 }
